pull animation position logic into separate util
